requirements.txt
Django==1.10
djangorestframework==3.7.1
elasticsearch==5.6.3
boto3==1.4.3
django-extentions==1.9.6